fix: use Menu for MenuSearchResults and separate MenuContent styles
