+++ src/components/Menu/Menu.scss
@@ -38,66 +38,14 @@ $g-c-menu-z: 60 !default;
 	visibility: visible;
 }
 
-.c-menu__item {
-	text-decoration: none;
-	color: inherit;
-	display: flex;
-	padding: 0.8rem 1.6rem;
-	position: relative;
-
-	&:hover {
-		background-color: $color-gray-50;
-	}
-
-	&.c-menu__item--large {
-		flex-direction: column;
-		white-space: normal;
-	}
-
-	&.c-menu__item--center {
-		align-items: center;
-	}
-}
-
-.c-menu__item--nowrap {
-	white-space: nowrap;
-}
-
-.c-menu__item--selected {
-	font-weight: 600;
-}
-
-.c-menu__label {
-	overflow: hidden;
-	text-overflow: ellipsis;
-
-	.o-svg-icon {
-		margin-right: 1rem;
-	}
-}
-
-.c-menu__divider {
-	border-top: 0.1rem solid $color-gray-200;
-	display: block;
-	margin-top: 0.4rem;
-	padding-top: 0.4rem;
-}
-
-
-.c-menu__icon {
-	opacity: 0.5;
-	margin-right: 0.5rem;
-
-	.c-menu__item:hover & {
-		opacity: 1;
-	}
-}
-
 .c-menu--search-result {
 	.c-menu__item {
 		justify-content: space-between;
 		padding: 1.2rem 1.6rem;
-		border-bottom: 0.1rem solid $color-gray-100;
+
+		&:not(:last-child) {
+			border-bottom: 0.1rem solid $color-gray-100;
+		}
 	}
 
 	.c-menu__footer {
